// src/app/models/openingStockFibreDts.ts
export interface OpeningStockFibreDts {
  receivedDCId: number;
  stockAddedDate: string;
  receivedDtsId: number;
  fiberTypeId: number;
  fiberTypeName: string;
  fiberShadeId: number;
  fiberShadeName: string;
  lot: string;
  hsnCode: number;
  stockWeight: number;
  stockBales: number;
  availableBalance: number;
}